Big change - rewrote collision detection from the ground up. Not sure why I was so hesitant to get deltas from sprites previous/current coordinates, but holy cow that was way easier and more accurate. Everything's behaving the way I want it to, with the exception of some minor hitbox calculations (I suspect, not 100% sure), but otherwise, big win.

+++ RetroWar/RetroWar/Services/Interfaces/Helpers/Model/IFaceHelper.cs
@@ -4,7 +4,7 @@ namespace RetroWar.Services.Interfaces.Helpers.Model
 {
     public interface IFaceHelper
     {
-        int GetFaceAxis(Sprite sprite, Face face);
-        int GetFaceDifference(Sprite normal, Face normalFace, Sprite based);
+        float GetFaceAxis(Sprite sprite, Face face);
+        float GetFaceDifference(Sprite normal, Face normalFace, Sprite based);
     }
 }
